Add read-only article query service
