+++ sentinela/application/services.py
@@ -8,7 +8,11 @@ import logging
 import time
 
 from sentinela.domain.entities import Article, Portal
-from sentinela.domain.repositories import ArticleRepository, PortalRepository
+from sentinela.domain.repositories import (
+    ArticleReadRepository,
+    ArticleRepository,
+    PortalRepository,
+)
 from sentinela.infrastructure.scraper import Scraper
 
 
@@ -99,6 +103,7 @@ class NewsCollectorService:
         start_dt = datetime.combine(start_date, datetime.min.time())
         end_dt = datetime.combine(end_date, datetime.max.time())
         return self._article_repository.list_by_period(portal_name, start_dt, end_dt)
+
 
     def collect_all_for_portal(
         self,
@@ -233,3 +238,17 @@ class NewsCollectorService:
             )
         )
         return all_new
+
+
+class ArticleQueryService:
+    """Provides read-only access to collected articles."""
+
+    def __init__(self, repository: ArticleReadRepository) -> None:
+        self._repository = repository
+
+    def list_articles(
+        self, portal_name: str, start_date: date, end_date: date
+    ) -> Iterable[Article]:
+        start_dt = datetime.combine(start_date, datetime.min.time())
+        end_dt = datetime.combine(end_date, datetime.max.time())
+        return self._repository.list_by_period(portal_name, start_dt, end_dt)
