+++ sentinela/services/publications/__init__.py
@@ -1,0 +1,5 @@
+"""Publications service dependency container."""
+
+from .container import PublicationsContainer, build_publications_container
+
+__all__ = ["PublicationsContainer", "build_publications_container"]
